Renamed `todo` to `lib`
